refactor(P0): remove dead CSV code, WriteQueue stub, rename csv_manager
- Remove merge_sqlite_to_csv() from result_aggregator.py (dead code)
- Remove orphan _build_csv_updates() from utils.py (dead code)
- Remove duplicate @staticmethod _build_csv_updates from main.py
- Remove WriteQueue stub class + update create_handler() factory to return
  (CSVHandler, str) and update all 3 callers (csv_manager, processing_controller,
  worker_manager) accordingly
- Rename csv_manager.py -> persistence_manager.py + update imports in
  main.py, processing_service.py, tests
- Delete legacy data/control/download_control*.csv files (no code reads them)
- Fix BrowserInitError to use kwargs.setdefault to allow context override

perf(P1/P3): reduce download delays
- Reduce hide_and_click and rapid-trigger sleep from 0.3s to 0.1s each
- OS-conditional finalize_folder cooldown: 1.0s (win32) vs 0.3s (others)
