+++ src/core/utils.py
@@ -139,48 +139,6 @@ def _suffix_for_status(status_code: str) -> str:
     return mapping.get(status_code, f'_{status_code}')
 
 
-
-
-
-
-def _build_csv_updates(result: Dict[str, Any]) -> Dict[str, Any]:
-    """Translate a processing result into CSV column updates."""
-    from datetime import datetime
-    status_code = (result.get('status_code') or '').upper() or 'FAILED'
-    attachment_names = result.get('attachment_names') or []
-    if isinstance(attachment_names, str):
-        attachment_names = [name for name in attachment_names.split(';') if name]
-
-    error_message = ''
-    success = result.get('success')
-    if success is None:
-        success = status_code in {'COMPLETED', 'NO_ATTACHMENTS', 'PARTIAL'}
-    if not success:
-        error_message = result.get('message', '') or result.get('error', '')
-
-    updates: Dict[str, Any] = {
-        'STATUS': status_code,
-        'ATTACHMENTS_FOUND': result.get('attachments_found', 0),
-        'ATTACHMENTS_DOWNLOADED': result.get('attachments_downloaded', 0),
-        'AttachmentName': attachment_names,
-        'DOWNLOAD_FOLDER': result.get('final_folder', ''),
-        'COUPA_URL': result.get('coupa_url', ''),
-        'ERROR_MESSAGE': error_message,
-    }
-
-    supplier_name = result.get('supplier_name')
-    if supplier_name:
-        updates['SUPPLIER'] = supplier_name
-
-    last_processed = result.get('last_processed')
-    if isinstance(last_processed, datetime):
-        updates['LAST_PROCESSED'] = last_processed.isoformat()
-    elif isinstance(last_processed, str) and last_processed:
-        updates['LAST_PROCESSED'] = last_processed
-
-    return updates
-
-
 def _compose_csv_message(result_payload: dict) -> str:
     """Compose error message for CSV from result payload."""
     status_code = (result_payload.get('status_code') or '').upper()
